refactor: move logic into context provider
